pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclawwatch"
version = "0.1.0"
description = "Local-first OTel-native observability for AI agents"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["ai", "agents", "observability", "opentelemetry", "llm"]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "tomli>=2.0; python_version < '3.11'",
    "tomli-w>=1.0",
    "duckdb>=0.10",
    "jsonschema>=4.0",
    "genson>=1.2",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp-proto-http>=1.25",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
    "opentelemetry-exporter-prometheus>=0.46b0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "websockets>=12.0",
]

[project.optional-dependencies]
langchain = ["langchain>=0.2"]
crewai    = ["crewai>=0.28"]
autogen   = ["pyautogen>=0.2"]
dev       = ["pytest", "pytest-asyncio", "httpx", "ruff", "mypy"]

[project.scripts]
ocw = "ocw.cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/synthetic", "tests/agents", "tests/integration"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["ocw"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
